svn path=/tags/KDE_2_2_2_RELEASE/kdelibs/; revision=184287

+++ kdoctools/xml2man.cpp
@@ -7,7 +7,7 @@
 #include <libxml/HTMLtree.h>
 #include <libxml/xmlIO.h>
 #include <libxml/parserInternals.h>
-#include <libxslt/xslt.h>
+#include <libxslt/xsltconfig.h>
 #include <libxslt/xsltInternals.h>
 #include <libxslt/transform.h>
 #include <libxslt/xsltutils.h>
@@ -16,6 +16,7 @@
 #include <kinstance.h>
 #include <xslt.h>
 #include <qfile.h>
+#include <qregexp.h>
 #include <kcmdlineargs.h>
 #include <klocale.h>
 #include <kaboutdata.h>
@@ -96,7 +97,7 @@ int main(int argc, char **argv) {
                 index = index2 + 2;
             }
         }
-        if ( index >= contents.length() )
+        if ( index >= (int)contents.length() )
             break;
     }
